requirements.txt
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
cryptography==2.8
decorator==4.4.1
Django==3.0.4
django-cors-headers==3.2.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-nine==0.2.2
djangorestframework==3.8.2
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
idna==2.8
ipython==7.9.0
ipython-genutils==0.2.0
jedi==0.15.1
jwt==0.6.1
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.5.2
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
six==1.13.0
sqlparse==0.3.1
traitlets==4.3.3
urllib3==1.25.7
virtualenv==16.7.8
wcwidth==0.1.7
